Adds migration utils
