+++ lib/migration.js
@@ -14,6 +14,17 @@ const { confirm, STATE_SUCCESS, STATE_FAILURE } = require('./config');
 
 const { storeMigration, getNewMigrations } = require('./backend');
 
+const migrationHeader = stripIndent`/* eslint-env node */
+  const { withUtils } = require('@jungvonmatt/contentful-migrations');
+
+  /**
+   * Contentful migration
+   * API: https://github.com/contentful/contentful-migration
+   * Editor Interfaces: https://www.contentful.com/developers/docs/extensibility/app-framework/editor-interfaces/
+   */
+  module.exports = withUtils(async function (migration, context, utils) {
+`
+
 /**
  * Create new migration file.
  * Adds initial migration file adding the migration field in the content type
@@ -32,16 +43,9 @@ const createMigration = async (config) => {
   const { directory } = config || {};
   const timestamp = Date.now();
   const filename = path.join(directory, `${timestamp}-migration.${module ? 'cjs' : 'js'}`);
-  const content = stripIndent`
-  /* eslint-env node */
-
-  /**
-   * Contentful migration
-   */
-  module.exports = function(migration /*, context */) {
+  const content = stripIndent`${migrationHeader}
     // Add your migration code here
-    // See: https://github.com/contentful/contentful-migration
-  }`;
+  })`;
 
   await fs.outputFile(filename, await format(filename, content));
   console.log(`Generated new migration file to ${chalk.green(filename)}`);
@@ -81,7 +85,23 @@ const fetchMigration = async (config) => {
 
     const content = await generateMigrationScript(client, [entry]);
 
-    await fs.outputFile(filename, await format(filename, content));
+    let modifiedContent = content.toString();
+    if (/(defaultValue\({\s*)[^\:]+([^\)]+)/g.test(modifiedContent)) {
+      modifiedContent = content.toString()
+      .replace('module.exports = function (migration) {', `${migrationHeader}
+        const defaultLocale = utils.getDefaultLocale();
+
+      `)
+      .replace(/};\s*$/g, '});')
+      .replace(/(defaultValue\({\s*)[^\:]+([^\)]+)/g, '$1[defaultLocale.code]$2');
+
+    } else {
+      modifiedContent = content.toString()
+      .replace('module.exports = function (migration) {', migrationHeader)
+      .replace(/};\s*$/g, '});');
+    }
+
+    await fs.outputFile(filename, await format(filename, modifiedContent));
     console.log(`Generated new migration file to ${chalk.green(filename)}`);
   });
